+++ lib/datac_annotations.dart
@@ -1,6 +1,13 @@
 library datac_annotations;
 
-export 'package:datac_annotations/src/data_class.dart';
-export 'package:datac_annotations/src/data_class_methods.dart';
-export 'package:datac_annotations/src/ignore.dart';
-export 'package:datac_annotations/src/qualifier.dart';
+export 'package:datac_annotations/src/annotations/data_class.dart';
+export 'package:datac_annotations/src/annotations/data_class_methods.dart';
+export 'package:datac_annotations/src/annotations/ignore.dart';
+export 'package:datac_annotations/src/annotations/qualifier.dart';
+export 'package:datac_annotations/src/annotations/equals.dart';
+export 'package:datac_annotations/src/annotations/list_equals.dart';
+export 'package:datac_annotations/src/operators/equals_evaluator.dart';
+export 'package:datac_annotations/src/operators/list_equals_evaluator.dart';
+export 'package:datac_annotations/src/operators/map_equals_evaluator.dart';
+export 'package:datac_annotations/src/operators/data_time_equals_evaluator.dart';
+export 'package:datac_annotations/src/annotations/date_equals.dart';