+++ lib/src/annotations/data_class_methods.dart
@@ -1,4 +1,5 @@
 /// an enumerate represent all method available.
+/// json - represent both (fromJson, toJson) methods.
 enum DataClassMethods{
   fromJson, toJson, tostring, equals, hashcode, json, copyWith
 }