+++ lib/src/annotations/list_equals.dart
@@ -1,0 +1,6 @@
+import 'package:datac_annotations/src/annotations/equals.dart';
+import 'package:datac_annotations/src/operators/list_equals_evaluator.dart';
+/// Field annotation, used to add equality operator to list field.
+class ListEquals extends Equals{
+  const ListEquals(): super(evaluator: ListEqualsEvaluator);
+}