+++ lib/src/operators/equals_evaluator.dart
@@ -1,0 +1,4 @@
+/// interface to create custom equality operator.
+abstract class EqualsEvaluator<T>{
+  bool evaluate(T obj, Object other);
+}