+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-## 0.0.5
-
+## 0.0.5 
+## 0.0.6
+## add support to custom equals evaluator and built-in equals evaluators:
+- 1 - List equals evaluator.
+- 2 - Map equals evaluator.
+- 3 - DateTime equals evaluator.